chapter-3/requirements.txt
discord.py
boto3
weasyprint